fix: address remaining clang-tidy warnings on PR #540
- LazyComplexVector.cpp: const-qualify auto* in copyRanges and decode
- LazyComplexVector.h: drop default arg on resize override (provided by base)
- LazyComplexCodec.cpp / LazyBundleEncoder.cpp: use auto for cast-init lines
  and uppercase the integer-literal suffix (1u -> 1U)
- ShuffleReaderNode.cpp: replace using-directive with explicit using-decls
- RowContainer.cpp: add (unknown) tag to TODO comment
- StreamingAggregation: undo my earlier static_cast<uint32_t> and align the
  groupNumberThreshold_ type with upstream (#538), which switched the field
  to vector_size_t. The merged tree now narrowed uint32_t->vector_size_t, so
  the cast became the wrong direction; matching the type removes the cast
  entirely and keeps local and merged states consistent.

+++ bolt/exec/StreamingAggregation.cpp
@@ -44,7 +44,7 @@ StreamingAggregation::StreamingAggregation(
               ? "PartialStreamingAggregation"
               : "StreamingAggregation"),
       outputBatchSize_{outputBatchRows()},
-      groupNumberThreshold_{static_cast<uint32_t>(2 * outputBatchSize_)},
+      groupNumberThreshold_{2 * outputBatchSize_},
       aggregationNode_{aggregationNode},
       step_{aggregationNode->step()} {
   if (aggregationNode_->ignoreNullKeys()) {
